+++ vscode/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to the "dependi" extension will be documented in this file.
 
 ## [v0.7.9](https://github.com/filllabs/dependi/compare/v0.7.8...v0.7.9)
 
+### New Features
+
+- Commands to toggle Lock File Parsing in the title bar.
+  - ![Lock Icon](https://www.dependi.io/screenshots/lock.png)  Lock File Parsing enabled.
+  - ![Unlock Icon](https://www.dependi.io/screenshots/unlock.png)  Lock File Parsing disabled.
+  - ![Doc-Lock Icon](https://www.dependi.io/screenshots/lock-doc.png)  A lock file has been found and parsed.
+
+
 ### Improvements
 
 - New advanced vulnerability report. (_Only for [Pro](https://dependi.io) version._)